backend/requirements.txt
django==5.2.13
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
pymysql==1.1.2
python-dotenv==1.2.2
